Add files via upload
Added comments next to sections I worked with for more info. If you have any questions just ask.

Line 8 & 9: Changed the character used for the player and split it between Player One (O) and Player Two (X). Was not able to figure out a way to do it in the middle of the game so now all it does is change the person's image.

Line 52 to 59: Put all the data needed for writing text and values needed for future code here. Number, Choices, and Options are used for the player choice part. po and px are used for checking that the player put in a correct number. P is used later to determine the character's image. X and Y are used as part of the wall subprogram to decide where it should start and end.

Line 91 to 110: Contains the code for checking which character the player wants to play and checks that the answer works.

Line 347 to 369: Subprograms for creating the walls. Uses X and Y to find out the positions and how long they should be.

Line 297 to 311: Used to find what the character should look like. Uses P for getting the number to decide between the two.